after user create from angular side some update here

+++ mmsystem/src/main/java/com/mmsystem/property/repo/MmsUserRoleJPARepository.java
@@ -3,6 +3,6 @@ package com.mmsystem.property.repo;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.mmsystem.property.model.MmsUserRole;
 
-public interface MmsUserRoleJPARepository extends JpaRepository<MmsUserRole, Long> {
+public interface MmsUserRoleJPARepository extends JpaRepository<MmsUserRole, Integer> {
 
 }
